MOD-16382 Consolidate topology tests into one file
Merge test_ts_topology_notifications.py into test_topology_changes.py -- one
topology-tests file instead of two. The cold-coordinator first-multishard-query
test moves in as a third section (CLUSTERSET-apply, auto-refresh/reshard,
cold-fan-out); imports already cover it (Env, VALGRIND, SANITIZER, redis). No
test logic changes; 11 test cases total.
